move ldap_user_search to m-f-d
